Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,12 +1,10 @@
 #include <stdio.h>
 int main(){
     int a,b,c;
-    float avg;
+    int avg;
     scanf("%d",&a);
     scanf("%d",&b);
     scanf("%d",&c);
     avg=a+b+c/3;
-    printf("Average: %.2f",avg);
-
-
+    printf("Average: %d",avg);
 }
